restructured files to be grouped related to their place in the component tree
