+++ client/src/components/StudentDetails/PracticePlan/PracticePlanCard.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useContext } from "react";
-import CreatePracticePlan from "../components/CreatePracticePlan";
-import PracticePlan from "../components/PracticePlan";
+import CreatePracticePlan from "./CreatePracticePlan";
+import PracticePlan from "./PracticePlan";
 import { Typography, Button } from "@mui/joy";
-import { StudentContext } from "../pages/StudentDetails";
+import { StudentContext } from "../../../pages/StudentDetails";
 
 // displays practice plans from a student
 export default function PracticePlanCard() {
